refactor(database): user repository service + test
+ split test data from test, needs adjusting tests
