+++ test/services/firebase/testing_auth_providers.dart
@@ -5,7 +5,7 @@ import "package:google_sign_in_mocks/google_sign_in_mocks.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/services/login_service.dart";
 
-import "../mock_data/firebase_auth_user_mock.dart";
+import "../test_data/firebase_auth_user_mock.dart";
 
 final firebaseAuthMocksOverrides = [
   googleSignInProvider.overrideWith(mockGoogleSignIn),
